+++ telegrambots-abilities/src/main/java/org/telegram/abilitybots/api/sender/DefaultMessageSender.java
@@ -70,67 +70,67 @@ public class DefaultMessageSender implements MessageSender {
 
   @Override
   public Boolean answerInlineQuery(AnswerInlineQuery answerInlineQuery) throws TelegramApiException {
-    return bot.answerInlineQuery(answerInlineQuery);
+    return bot.execute(answerInlineQuery);
   }
 
   @Override
   public Boolean sendChatAction(SendChatAction sendChatAction) throws TelegramApiException {
-    return bot.sendChatAction(sendChatAction);
+    return bot.execute(sendChatAction);
   }
 
   @Override
   public Message forwardMessage(ForwardMessage forwardMessage) throws TelegramApiException {
-    return bot.forwardMessage(forwardMessage);
+    return bot.execute(forwardMessage);
   }
 
   @Override
   public Message sendLocation(SendLocation sendLocation) throws TelegramApiException {
-    return bot.sendLocation(sendLocation);
+    return bot.execute(sendLocation);
   }
 
   @Override
   public Message sendVenue(SendVenue sendVenue) throws TelegramApiException {
-    return bot.sendVenue(sendVenue);
+    return bot.execute(sendVenue);
   }
 
   @Override
   public Message sendContact(SendContact sendContact) throws TelegramApiException {
-    return bot.sendContact(sendContact);
+    return bot.execute(sendContact);
   }
 
   @Override
   public Boolean kickMember(KickChatMember kickChatMember) throws TelegramApiException {
-    return bot.kickMember(kickChatMember);
+    return bot.execute(kickChatMember);
   }
 
   @Override
   public Boolean unbanMember(UnbanChatMember unbanChatMember) throws TelegramApiException {
-    return bot.unbanMember(unbanChatMember);
+    return bot.execute(unbanChatMember);
   }
 
   @Override
   public Boolean leaveChat(LeaveChat leaveChat) throws TelegramApiException {
-    return bot.leaveChat(leaveChat);
+    return bot.execute(leaveChat);
   }
 
   @Override
   public Chat getChat(GetChat getChat) throws TelegramApiException {
-    return bot.getChat(getChat);
+    return bot.execute(getChat);
   }
 
   @Override
   public List<ChatMember> getChatAdministrators(GetChatAdministrators getChatAdministrators) throws TelegramApiException {
-    return bot.getChatAdministrators(getChatAdministrators);
+    return bot.execute(getChatAdministrators);
   }
 
   @Override
   public ChatMember getChatMember(GetChatMember getChatMember) throws TelegramApiException {
-    return bot.getChatMember(getChatMember);
+    return bot.execute(getChatMember);
   }
 
   @Override
   public Integer getChatMemberCount(GetChatMemberCount getChatMemberCount) throws TelegramApiException {
-    return bot.getChatMemberCount(getChatMemberCount);
+    return bot.execute(getChatMemberCount);
   }
 
   @Override
@@ -140,117 +140,117 @@ public class DefaultMessageSender implements MessageSender {
 
   @Override
   public Boolean deleteChatPhoto(DeleteChatPhoto deleteChatPhoto) throws TelegramApiException {
-    return bot.deleteChatPhoto(deleteChatPhoto);
+    return bot.execute(deleteChatPhoto);
   }
 
   @Override
   public void deleteChatPhoto(DeleteChatPhoto deleteChatPhoto, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.deleteChatPhoto(deleteChatPhoto, sentCallback);
+    bot.executeAsync(deleteChatPhoto, sentCallback);
   }
 
   @Override
   public Boolean pinChatMessage(PinChatMessage pinChatMessage) throws TelegramApiException {
-    return bot.pinChatMessage(pinChatMessage);
+    return bot.execute(pinChatMessage);
   }
 
   @Override
   public void pinChatMessage(PinChatMessage pinChatMessage, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.pinChatMessage(pinChatMessage, sentCallback);
+    bot.executeAsync(pinChatMessage, sentCallback);
   }
 
   @Override
   public Boolean unpinChatMessage(UnpinChatMessage unpinChatMessage) throws TelegramApiException {
-    return bot.unpinChatMessage(unpinChatMessage);
+    return bot.execute(unpinChatMessage);
   }
 
   @Override
   public void unpinChatMessage(UnpinChatMessage unpinChatMessage, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.unpinChatMessage(unpinChatMessage, sentCallback);
+    bot.executeAsync(unpinChatMessage, sentCallback);
   }
 
   @Override
   public Boolean promoteChatMember(PromoteChatMember promoteChatMember) throws TelegramApiException {
-    return bot.promoteChatMember(promoteChatMember);
+    return bot.execute(promoteChatMember);
   }
 
   @Override
   public void promoteChatMember(PromoteChatMember promoteChatMember, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.promoteChatMember(promoteChatMember, sentCallback);
+    bot.executeAsync(promoteChatMember, sentCallback);
   }
 
   @Override
   public Boolean restrictChatMember(RestrictChatMember restrictChatMember) throws TelegramApiException {
-    return bot.restrictChatMember(restrictChatMember);
+    return bot.execute(restrictChatMember);
   }
 
   @Override
   public void restrictChatMember(RestrictChatMember restrictChatMember, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.restrictChatMember(restrictChatMember, sentCallback);
+    bot.executeAsync(restrictChatMember, sentCallback);
   }
 
   @Override
   public Boolean setChatDescription(SetChatDescription setChatDescription) throws TelegramApiException {
-    return bot.setChatDescription(setChatDescription);
+    return bot.execute(setChatDescription);
   }
 
   @Override
   public void setChatDescription(SetChatDescription setChatDescription, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.setChatDescription(setChatDescription, sentCallback);
+    bot.executeAsync(setChatDescription, sentCallback);
   }
 
   @Override
   public Boolean setChatTite(SetChatTitle setChatTitle) throws TelegramApiException {
-    return bot.setChatTitle(setChatTitle);
+    return bot.execute(setChatTitle);
   }
 
   @Override
   public void setChatTite(SetChatTitle setChatTitle, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.setChatTitle(setChatTitle, sentCallback);
+    bot.executeAsync(setChatTitle, sentCallback);
   }
 
   @Override
   public String exportChatInviteLink(ExportChatInviteLink exportChatInviteLink) throws TelegramApiException {
-    return bot.exportChatInviteLink(exportChatInviteLink);
+    return bot.execute(exportChatInviteLink);
   }
 
   @Override
   public void exportChatInviteLinkAsync(ExportChatInviteLink exportChatInviteLink, SentCallback<String> sentCallback) throws TelegramApiException {
-    bot.exportChatInviteLinkAsync(exportChatInviteLink, sentCallback);
+    bot.executeAsync(exportChatInviteLink, sentCallback);
   }
 
   @Override
   public Boolean deleteMessage(DeleteMessage deleteMessage) throws TelegramApiException {
-    return bot.deleteMessage(deleteMessage);
+    return bot.execute(deleteMessage);
   }
 
   @Override
   public void deleteMessageAsync(DeleteMessage deleteMessage, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.deleteMessage(deleteMessage, sentCallback);
+    bot.executeAsync(deleteMessage, sentCallback);
   }
 
   @Override
   public Serializable editMessageText(EditMessageText editMessageText) throws TelegramApiException {
-    return bot.editMessageText(editMessageText);
+    return bot.execute(editMessageText);
   }
 
   @Override
   public Serializable editMessageCaption(EditMessageCaption editMessageCaption) throws TelegramApiException {
-    return bot.editMessageCaption(editMessageCaption);
+    return bot.execute(editMessageCaption);
   }
 
   @Override
   public Serializable editMessageReplyMarkup(EditMessageReplyMarkup editMessageReplyMarkup) throws TelegramApiException {
-    return bot.editMessageReplyMarkup(editMessageReplyMarkup);
+    return bot.execute(editMessageReplyMarkup);
   }
 
   @Override
   public Boolean answerCallbackQuery(AnswerCallbackQuery answerCallbackQuery) throws TelegramApiException {
-    return bot.answerCallbackQuery(answerCallbackQuery);
+    return bot.execute(answerCallbackQuery);
   }
 
   @Override
   public UserProfilePhotos getUserProfilePhotos(GetUserProfilePhotos getUserProfilePhotos) throws TelegramApiException {
-    return bot.getUserProfilePhotos(getUserProfilePhotos);
+    return bot.execute(getUserProfilePhotos);
   }
 
   @Override
@@ -275,7 +275,7 @@ public class DefaultMessageSender implements MessageSender {
 
   @Override
   public File getFile(GetFile getFile) throws TelegramApiException {
-    return bot.getFile(getFile);
+    return bot.execute(getFile);
   }
 
   @Override
@@ -290,127 +290,127 @@ public class DefaultMessageSender implements MessageSender {
 
   @Override
   public Serializable setGameScore(SetGameScore setGameScore) throws TelegramApiException {
-    return bot.setGameScore(setGameScore);
+    return bot.execute(setGameScore);
   }
 
   @Override
   public Serializable getGameHighScores(GetGameHighScores getGameHighScores) throws TelegramApiException {
-    return bot.getGameHighScores(getGameHighScores);
+    return bot.execute(getGameHighScores);
   }
 
   @Override
   public Message sendGame(SendGame sendGame) throws TelegramApiException {
-    return bot.sendGame(sendGame);
+    return bot.execute(sendGame);
   }
 
   @Override
   public Boolean deleteWebhook(DeleteWebhook deleteWebhook) throws TelegramApiException {
-    return bot.deleteWebhook(deleteWebhook);
+    return bot.execute(deleteWebhook);
   }
 
   @Override
   public Message sendMessage(SendMessage sendMessage) throws TelegramApiException {
-    return bot.sendMessage(sendMessage);
+    return bot.execute(sendMessage);
   }
 
   @Override
   public void sendMessageAsync(SendMessage sendMessage, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.sendMessageAsync(sendMessage, sentCallback);
+    bot.executeAsync(sendMessage, sentCallback);
   }
 
   @Override
   public void answerInlineQueryAsync(AnswerInlineQuery answerInlineQuery, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.answerInlineQueryAsync(answerInlineQuery, sentCallback);
+    bot.executeAsync(answerInlineQuery, sentCallback);
   }
 
   @Override
   public void sendChatActionAsync(SendChatAction sendChatAction, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.sendChatActionAsync(sendChatAction, sentCallback);
+    bot.executeAsync(sendChatAction, sentCallback);
   }
 
   @Override
   public void forwardMessageAsync(ForwardMessage forwardMessage, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.forwardMessageAsync(forwardMessage, sentCallback);
+    bot.executeAsync(forwardMessage, sentCallback);
   }
 
   @Override
   public void sendLocationAsync(SendLocation sendLocation, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.sendLocationAsync(sendLocation, sentCallback);
+    bot.executeAsync(sendLocation, sentCallback);
   }
 
   @Override
   public void sendVenueAsync(SendVenue sendVenue, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.sendVenueAsync(sendVenue, sentCallback);
+    bot.executeAsync(sendVenue, sentCallback);
   }
 
   @Override
   public void sendContactAsync(SendContact sendContact, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.sendContactAsync(sendContact, sentCallback);
+    bot.executeAsync(sendContact, sentCallback);
   }
 
   @Override
   public void kickMemberAsync(KickChatMember kickChatMember, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.kickMemberAsync(kickChatMember, sentCallback);
+    bot.executeAsync(kickChatMember, sentCallback);
   }
 
   @Override
   public void unbanMemberAsync(UnbanChatMember unbanChatMember, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.unbanMemberAsync(unbanChatMember, sentCallback);
+    bot.executeAsync(unbanChatMember, sentCallback);
   }
 
   @Override
   public void leaveChatAsync(LeaveChat leaveChat, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.leaveChatAsync(leaveChat, sentCallback);
+    bot.executeAsync(leaveChat, sentCallback);
   }
 
   @Override
   public void getChatAsync(GetChat getChat, SentCallback<Chat> sentCallback) throws TelegramApiException {
-    bot.getChatAsync(getChat, sentCallback);
+    bot.executeAsync(getChat, sentCallback);
   }
 
   @Override
   public void getChatAdministratorsAsync(GetChatAdministrators getChatAdministrators, SentCallback<ArrayList<ChatMember>> sentCallback) throws TelegramApiException {
-    bot.getChatAdministratorsAsync(getChatAdministrators, sentCallback);
+    bot.executeAsync(getChatAdministrators, sentCallback);
   }
 
   @Override
   public void getChatMemberAsync(GetChatMember getChatMember, SentCallback<ChatMember> sentCallback) throws TelegramApiException {
-    bot.getChatMemberAsync(getChatMember, sentCallback);
+    bot.executeAsync(getChatMember, sentCallback);
   }
 
   @Override
   public void getChatMemberCountAsync(GetChatMemberCount getChatMemberCount, SentCallback<Integer> sentCallback) throws TelegramApiException {
-    bot.getChatMemberCountAsync(getChatMemberCount, sentCallback);
+    bot.executeAsync(getChatMemberCount, sentCallback);
   }
 
   @Override
   public void editMessageTextAsync(EditMessageText editMessageText, SentCallback<Serializable> sentCallback) throws TelegramApiException {
-    bot.editMessageTextAsync(editMessageText, sentCallback);
+    bot.executeAsync(editMessageText, sentCallback);
   }
 
   @Override
   public void editMessageCaptionAsync(EditMessageCaption editMessageCaption, SentCallback<Serializable> sentCallback) throws TelegramApiException {
-    bot.editMessageCaptionAsync(editMessageCaption, sentCallback);
+    bot.executeAsync(editMessageCaption, sentCallback);
   }
 
   @Override
   public void editMessageReplyMarkup(EditMessageReplyMarkup editMessageReplyMarkup, SentCallback<Serializable> sentCallback) throws TelegramApiException {
-    bot.editMessageReplyMarkup(editMessageReplyMarkup, sentCallback);
+    bot.executeAsync(editMessageReplyMarkup, sentCallback);
   }
 
   @Override
   public void answerCallbackQueryAsync(AnswerCallbackQuery answerCallbackQuery, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.answerCallbackQueryAsync(answerCallbackQuery, sentCallback);
+    bot.executeAsync(answerCallbackQuery, sentCallback);
   }
 
   @Override
   public void getUserProfilePhotosAsync(GetUserProfilePhotos getUserProfilePhotos, SentCallback<UserProfilePhotos> sentCallback) throws TelegramApiException {
-    bot.getUserProfilePhotosAsync(getUserProfilePhotos, sentCallback);
+    bot.executeAsync(getUserProfilePhotos, sentCallback);
   }
 
   @Override
   public void getFileAsync(GetFile getFile, SentCallback<File> sentCallback) throws TelegramApiException {
-    bot.getFileAsync(getFile, sentCallback);
+    bot.executeAsync(getFile, sentCallback);
   }
 
   @Override
@@ -425,22 +425,22 @@ public class DefaultMessageSender implements MessageSender {
 
   @Override
   public void setGameScoreAsync(SetGameScore setGameScore, SentCallback<Serializable> sentCallback) throws TelegramApiException {
-    bot.setGameScoreAsync(setGameScore, sentCallback);
+    bot.executeAsync(setGameScore, sentCallback);
   }
 
   @Override
   public void getGameHighScoresAsync(GetGameHighScores getGameHighScores, SentCallback<ArrayList<GameHighScore>> sentCallback) throws TelegramApiException {
-    bot.getGameHighScoresAsync(getGameHighScores, sentCallback);
+    bot.executeAsync(getGameHighScores, sentCallback);
   }
 
   @Override
   public void sendGameAsync(SendGame sendGame, SentCallback<Message> sentCallback) throws TelegramApiException {
-    bot.sendGameAsync(sendGame, sentCallback);
+    bot.executeAsync(sendGame, sentCallback);
   }
 
   @Override
   public void deleteWebhook(DeleteWebhook deleteWebhook, SentCallback<Boolean> sentCallback) throws TelegramApiException {
-    bot.deleteWebhook(deleteWebhook, sentCallback);
+    bot.executeAsync(deleteWebhook, sentCallback);
   }
 
   @Override